% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Serial_gene_features_extraction.R
\name{Calc_feats}
\alias{Calc_feats}
\title{Calculate features of a sequence}
\usage{
Calc_feats(
  seq,
  PFAM_PATH = "databases/PFAM/Pfam-A.hmm",
  LAMBDA = 30,
  OMEGA = 0.05
)
}
\arguments{
\item{seq}{nucleotide sequence, works using a SeqFastadna object from the seqinr package.}

\item{PFAM_PATH}{Path to the Pfam-A database}

\item{LAMBDA}{Pseudo AA composition parameter}

\item{OMEGA}{Pseudo AA composition parameter}
}
\description{
Extract the features from a single nucleotide sequence. Also translate in the 3+ frames seaching for the longest ORF, using it to extract the features for the amino acid sequence.
}
