% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract_features.R
\name{Extract}
\alias{Extract}
\title{A feature extractor function.}
\usage{
Extract(
  FASTA_PATH = "",
  AAfile = "",
  PFAM_path = "",
  LAMBDA = 50,
  OMEGA = 0.05,
  CPU = 2,
  nuc_only = F,
  varGibbs_Model_path = "/mnt/DATABASES/bin/VarGibbs-2.2/data/AOP-CMB.par",
  subcel = F
)
}
\arguments{
\item{FASTA_PATH}{Path to the DNA fasta file.}

\item{PFAM_path}{Path to the PfamA database.}

\item{LAMBDA}{Pseudo amino acid composition parameter. Sequences with less amino acids than LAMBDA will be removed (50).}

\item{OMEGA}{Pseudo amino acid composition parameter (0.05).}

\item{CPU}{Number of threads to use. Each sequence will use one thread (2).}

\item{subcel}{use DeepLoc1.0 to predict probabilities for subcellular localization (F)}
}
\value{
A data frame of features
}
\description{
Input a DNA fasta file and it will output a vector of features for each sequence.
}
