% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Join_both.R
\name{Join_datasets}
\alias{Join_datasets}
\title{Join features from two datasets, changing the Class to E and NE}
\usage{
Join_datasets(essential = data.frame(), nonessential = data.frame())
}
\description{
Join features from two datasets, changing the Class to E and NE
}
