% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproduce_bacteria_results.R
\name{reproduce_bacteria_results}
\alias{reproduce_bacteria_results}
\title{Extract features using the procariote data from the package, then do a complete train and test to reproduces the results from the chapter 1.}
\usage{
reproduce_bacteria_results(CPU = 10)
}
\arguments{
\item{CPU}{number of threads to use, more threads uses more RAM and is faster.}
}
\description{
Extract features using the procariote data from the package, then do a complete train and test to reproduces the results from the chapter 1.
}
