% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproduce_insect_results.R
\name{reproduce_insect_results}
\alias{reproduce_insect_results}
\title{Reproduce Drosophila melanogaster and Tribolium castaneum train and tests}
\usage{
reproduce_insect_results(
  set1 = GeneEssentiality::drosophila_features,
  set2 = GeneEssentiality::tribolium_features,
  CPU = 20,
  seeds = GeneEssentiality::seed,
  plot_prefix = "Insects_",
  set1.name = "Dmel",
  set2.name = "Trib"
)
}
\arguments{
\item{set1}{Data.frame with the features of the first set. This will be used to train a model and as a test for the model trained from the second set. Default is the Drosophila melanogaster features}

\item{set2}{Data.frame with the features of the second set. This will be used to train a model and as a test for the model trained from the first set. Default is the Tribolium castaneum features}

\item{CPU}{Number of threads to use}

\item{seeds}{List of vectors, 30 lists of vectors with 6 elements and the last list with a single number, for the final model}

\item{plot_prefix}{Input a prefix for the filename of the plots. Default is "Insects_"}

\item{set1.name}{Name to be used for the title of the plots for the first set. Default is "Dmel"}

\item{set2.name}{Name to be used for the title of the plots for the first set. Default is "Trib"}
}
\description{
Run a complete train and test using the data from the package. Reproduces the paper results
Can also be used to compare two datasets with the same comparsion made for the insects. The required column is the Clas, and its values must be "E" or "NE".
Reditect to an object to save the models, ROCs results and Pvalues from the De longs test.  Plots are outputed to a file in the working dir.
}
