% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract_and_train.R
\name{RUN_all}
\alias{RUN_all}
\title{Run a complete train using data}
\usage{
RUN_all(
  PFAM_path = "/mnt/DATABASES/PFAM/Pfam-A.hmm",
  GE_PATH = system.file("extdata", "GE.fasta", package = "GeneEssentiality"),
  GNE_PATH = system.file("extdata", "GNE.fasta", package = "GeneEssentiality"),
  LAMBDA = 50,
  OMEGA = 0.05,
  CPU = 1,
  trees = 10,
  CV = 10,
  repeats = 3
)
}
\description{
Run a complete train using data
}
