% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Select_features.R
\name{Select}
\alias{Select}
\title{Remove features of little importance}
\usage{
Select(data = Complete_set)
}
\arguments{
\item{data}{A feature data frame including the Class label}
}
\value{
A reduced set of features
}
\description{
Remove features with zero variance and logical values that occur at a single point.
}
