% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{train_rf}
\alias{train_rf}
\title{Generate six RF models from two classes}
\usage{
train_rf(
  features = Complete_set,
  CPU = 2,
  trees = 1000,
  CV = 10,
  nrepeats = 3,
  seeds = seed,
  saveprediction = "final"
)
}
\arguments{
\item{CPU}{Number of threads to use when training the model}

\item{trees}{Number of trees for the random forest (rf) model}

\item{CV}{Number of sets for the cross-validation}

\item{nrepeats}{}

\item{data}{A feature data frame including the label of two Classes}
}
\value{
A list of models
}
\description{
This returns a list with one random forest model
}
\details{
The followind had no significant improvement and took too much time. In the future it may be activated with another parameter.
\emph{It will return a list with 3 normal rf models and 3 rf models made with randomized labels.
The models 1 and 2 are trained without any subsampling.
#' The models 3 and 4 are trained using upsampling during the cross-validation (Subsampling During Resampling).
The models 5 and 6 are trained using downsampling during the cross-validation (Subsampling During Resampling).}
}
