% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_models_with_datasets.R
\name{VS_models_plot}
\alias{VS_models_plot}
\title{Test using the provided models and features then plot the ROC and PRROC.
Used to compare two datasets with the same comparsion made for the insects, leaving one dataset out. The required column is the Class, and its values must be "E" or "NE".
Reditect to an object to save the models, ROCs results and Pvalues.  Plots are outputed to a file in the working dir.}
\usage{
VS_models_plot(
  set1_model = list(),
  set2_model = list(),
  set1_data = "",
  set2_data = "",
  file_prefix = "Compared_",
  set1.dataname = "Set1",
  set2.dataname = "Set2",
  set1.modelname = "set1_model",
  set2.modelname = "set2_model",
  test_vs_ZR = T
)
}
\arguments{
\item{file_prefix}{Input a prefix for the filename of the plots. Default is "Insects_"}

\item{test_vs_ZR}{Use pROC::roc.test of the tested ROC-AUC against a model that classifies every sample as the same class with the same probability (Zero Rule: ROC-AUC of 1.5). If set to FALSE, then test the results in a all vs all. If there is a single model and test_vs_ZR = FALSE, it will revert to TRUE. If set to "both", show both.}

\item{set1}{list of models. Each model is an object given by the train function of the caret package.}

\item{set2}{list of models. Each model is an object given by the train function of the caret package.}

\item{set1.name}{Name to be used for the title of the plots for the first set. Default is "Set1"}

\item{set2.name}{Name to be used for the title of the plots for the first set. Default is "Set2"}
}
\value{
The AUCs of the tests
}
\description{
Test using the provided models and features then plot the ROC and PRROC.
Used to compare two datasets with the same comparsion made for the insects, leaving one dataset out. The required column is the Class, and its values must be "E" or "NE".
Reditect to an object to save the models, ROCs results and Pvalues.  Plots are outputed to a file in the working dir.
}
