% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract_features.R
\name{Extract}
\alias{Extract}
\title{A feature extractor function.}
\usage{
Extract(
  PFAM_path = "/home/programs/DATABASES/PFAM/Pfam-A.hmm",
 
    FASTA_PATH = "/mnt/DATABASES/essenciais/drosophila/fastas/essential_complete-curated.fasta",
  LAMBDA = 50,
  OMEGA = 0.05,
  CPU = 2
)
}
\arguments{
\item{PFAM_path}{Path to the PfamA database.}

\item{FASTA_PATH}{Path to the DNA fasta file.}

\item{LAMBDA}{Pseudo amino acid composition parameter. Sequences with less amino acids than LAMBDA will be removed (50).}

\item{OMEGA}{Pseudo amino acid composition parameter (0.05).}

\item{CPU}{Number of threads to use. Each sequence will use one thread (2).}
}
\value{
A data frame of features
}
\description{
Input a DNA fasta file and it will output a vector of features for each sequence.
}
